agent/mcp_tools.py
"""
MCP Tools Integration for Deep Agent
Connects to CoinGecko, Firecrawl, and Santiment MCP servers
"""

import asyncio
import json
import subprocess
from typing import Any, Dict, List, Optional
from langchain_core.tools import BaseTool
from pydantic import BaseModel, Field


class MCPToolInput(BaseModel):
    """Input schema for MCP tools"""
    params: Dict[str, Any] = Field(description="Parameters for the MCP tool")


class MCPServerManager:
    """Manages connections to multiple MCP servers"""
    
    def __init__(self, config_path: str = "mcp_config.json"):
        with open(config_path, 'r') as f:
            config = json.load(f)
        self.servers = config.get("mcpServers", {})
        self.processes = {}
    
    async def start_servers(self):
        """Start all MCP servers"""
        for name, config in self.servers.items():
            try:
                print(f"Starting MCP server: {name}")
                # In production, you'd start these as background processes
                # For now, we'll just validate the configuration
                print(f"✓ {name} configured: {config['command']} {' '.join(config['args'])}")
            except Exception as e:
                print(f"✗ Failed to start {name}: {e}")
    
    async def stop_servers(self):
        """Stop all MCP servers"""
        for name, process in self.processes.items():
            if process:
                process.terminate()
                print(f"Stopped MCP server: {name}")


class CoinGeckoMCPTool(BaseTool):
    """Tool to query CoinGecko data via MCP"""
    
    name: str = "coingecko_query"
    description: str = """
    Query cryptocurrency data from CoinGecko Pro API.
    Use this to get:
    - Current prices of cryptocurrencies
    - Market data (volume, market cap, etc.)
    - Historical price data
    - Trending coins
    - Market trends
    
    Example params:
    - {"action": "get_price", "coin_id": "bitcoin", "currency": "usd"}
    - {"action": "get_trending"}
    - {"action": "get_market_data", "coin_id": "ethereum"}
    """
    
    async def _arun(self, params: Dict[str, Any]) -> str:
        """Execute CoinGecko query asynchronously"""
        try:
            # In production, this would call the actual MCP server
            # For dev mode, we'll simulate the response
            action = params.get("action", "get_price")
            
            if action == "get_price":
                coin_id = params.get("coin_id", "bitcoin")
                return f"Price data for {coin_id}: Currently fetching from CoinGecko API..."
            
            elif action == "get_trending":
                return "Trending coins: Fetching from CoinGecko API..."
            
            return f"Executing CoinGecko action: {action}"
        
        except Exception as e:
            return f"Error querying CoinGecko: {str(e)}"
    
    def _run(self, params: Dict[str, Any]) -> str:
        """Synchronous version"""
        return asyncio.run(self._arun(params))


class FirecrawlMCPTool(BaseTool):
    """Tool to scrape web content via Firecrawl MCP"""
    
    name: str = "firecrawl_scrape"
    description: str = """
    Scrape and extract content from websites using Firecrawl.
    Use this to:
    - Extract clean text from web pages
    - Get structured data from websites
    - Monitor crypto news sites
    - Gather market sentiment from blogs/news
    
    Example params:
    - {"url": "https://example.com", "format": "markdown"}
    - {"url": "https://cryptonews.com/article", "extract": ["title", "content"]}
    """
    
    async def _arun(self, params: Dict[str, Any]) -> str:
        """Execute Firecrawl scraping asynchronously"""
        try:
            url = params.get("url")
            if not url:
                return "Error: URL is required"
            
            return f"Scraping content from {url} via Firecrawl..."
        
        except Exception as e:
            return f"Error scraping with Firecrawl: {str(e)}"
    
    def _run(self, params: Dict[str, Any]) -> str:
        """Synchronous version"""
        return asyncio.run(self._arun(params))


class SantimentMCPTool(BaseTool):
    """Tool to query Santiment data via custom MCP"""
    
    name: str = "santiment_query"
    description: str = """
    Query Santiment API for advanced crypto market intelligence.
    Use this to get:
    - Social sentiment metrics
    - On-chain metrics
    - Development activity
    - Network growth
    - Token velocity and circulation
    
    Example params:
    - {"action": "get_sentiment", "coin": "bitcoin"}
    - {"action": "get_onchain_metrics", "coin": "ethereum", "metric": "active_addresses"}
    - {"action": "get_dev_activity", "project": "polkadot"}
    """
    
    async def _arun(self, params: Dict[str, Any]) -> str:
        """Execute Santiment query asynchronously"""
        try:
            action = params.get("action", "get_sentiment")
            coin = params.get("coin", "bitcoin")
            
            return f"Querying Santiment for {action} on {coin}..."
        
        except Exception as e:
            return f"Error querying Santiment: {str(e)}"
    
    def _run(self, params: Dict[str, Any]) -> str:
        """Synchronous version"""
        return asyncio.run(self._arun(params))


def get_mcp_tools() -> List[BaseTool]:
    """Get all MCP tools for the agent"""
    return [
        CoinGeckoMCPTool(),
        FirecrawlMCPTool(),
        SantimentMCPTool(),
    ]

