requirements.txt
fastapi
uvicorn
python-dotenv
deepagents
langchain-openai
tiktoken
psycopg2-binary
sqlalchemy[asyncio]
alembic
asyncpg
